+++ src/main/kotlin/minesweeper/dto/GameResultDto.kt
@@ -1,6 +1,11 @@
-package minesweeper.model
+package minesweeper.dto
 
-class GameResult private constructor(
+import minesweeper.model.Cell
+import minesweeper.model.Cells
+import minesweeper.model.GameStatus
+import minesweeper.model.MineBoard
+
+class GameResultDto private constructor(
     val boardRows: List<List<String>>,
     val gameStatus: GameStatus
 ) {
@@ -18,20 +23,20 @@ class GameResult private constructor(
         private const val MINE_MARK = "*"
         private const val CLOSE_MARK = "C"
 
-        fun from(board: MineBoard): GameResult {
+        fun from(board: MineBoard): GameResultDto {
             val boardResult = board.board.map { generateBoardRow(it, board) }
             val boardRows = boardResult.flatten()
 
             if (boardRows.contains(MINE_MARK)) {
-                return GameResult(boardResult, GameStatus.LOST)
+                return GameResultDto(boardResult, GameStatus.LOST)
             }
 
             val closeMarkCount = boardRows.count { it == CLOSE_MARK }
             if (board.mineCount == closeMarkCount) {
-                return GameResult(boardResult, GameStatus.WIN)
+                return GameResultDto(boardResult, GameStatus.WIN)
             }
 
-            return GameResult(boardResult, GameStatus.ONGOING)
+            return GameResultDto(boardResult, GameStatus.ONGOING)
         }
 
         private fun generateBoardRow(cells: Cells, board: MineBoard): List<String> =
