+++ src/test/kotlin/minesweeper/model/MineBoardTest.kt
@@ -1,5 +1,6 @@
 package minesweeper.model
 
+import minesweeper.dto.MineBoardCreateDto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
